MidiMessage now used a ByteArray internally instead of an IntArray, this makes things more uniform and predictable but means that the data in the array may not make much sense sometimes even if the end result is correct, ie, sending (144, 140, 130) will need you to do a toByte() call or java cast meaning what's sent (and received) is not what may have been intended but is technically correct because RtMidi will handle it correctly. I need to test this in action myself and may change this, but it seems that this works and is safe albeit a little vague

+++ src/main/kotlin/dev/basshelal/jnartmidi/api/ReadableMidiPort.kt
@@ -75,7 +75,7 @@ public class ReadableMidiPort : MidiPort<RtMidiInPtr> {
                 midiMessage?.also { midiMessage ->
                     // rare but necessary memalloc in RealTimeCritical code
                     if (midiMessage.size < size) midiMessage.size = size
-                    for (i in 0 until size) midiMessage[i] = message.getByte(i.toLong()).toInt()
+                    for (i in 0 until size) midiMessage[i] = message.getByte(i.toLong())
                     midiMessageCallback?.onMessage(midiMessage, timeStamp)
                 }
             } // end RealTimeCritical
